+++ src/core/render/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.100 (2020-11-17)
+
+#### :rocket: New Feature
+
+* Added support of filters with promises
+
 ## v3.0.0-rc.37 (2020-07-20)
 
 #### :house: Internal